refactor(actionpack): Wave 4 — testing relocation to action-dispatch/testing/
Mirrors Rails' `actionpack/lib/action_dispatch/testing/` layout:

- `action-controller/integration-test.ts` → `action-dispatch/testing/integration.ts`
  (with matching `.test.ts`). The class is `ActionDispatch::IntegrationTest`
  in Rails, so action-dispatch is the right home.
- Extract `TestRequest` → `action-dispatch/testing/test-request.ts`
  (Rails `test_request.rb`).
- Extract `LiveTestResponse` → `action-dispatch/testing/test-response.ts`
  (Rails `test_response.rb`).
- Extract `TestSession` → `action-dispatch/testing/test-process.ts`
  (Rails `test_process.rb`).
- `action-controller/test-case.ts` keeps the `TestCase` class only.
- Re-exports in `action-controller/index.ts` now reach into
  `action-dispatch/testing/` so the public surface is unchanged.

Deviation from plan: the assertions split (response/routing) is
deferred — `action-controller/template-assertions.ts` only contains
`assertTemplate`, and response/routing assertions live inline on
`TestCase` / `IntegrationTest`. Extracting them is a behavior-touching
refactor better done in its own PR.

api:compare actiondispatch 244→272 methods matched (+28). Overall
+25 net. test:compare totals unchanged. All 1793 actionpack tests
pass.
