+++ packages/actionpack/src/action-controller/test-case.ts
@@ -358,57 +358,6 @@ export class TestCase {
   }
 }
 
-export class TestRequest extends Request {
-  constructor(env: import("@blazetrails/rack").RackEnv = {}) {
-    super({
-      REQUEST_METHOD: "GET",
-      PATH_INFO: "/",
-      HTTP_HOST: "test.host",
-      SERVER_NAME: "test.host",
-      SERVER_PORT: "80",
-      ...env,
-    });
-  }
-}
-
-export class LiveTestResponse extends Response {}
-
-export class TestSession {
-  private _data = new Map<string, unknown>();
-
-  get(key: string): unknown {
-    return this._data.get(key);
-  }
-
-  set(key: string, value: unknown): void {
-    this._data.set(key, value);
-  }
-
-  has(key: string): boolean {
-    return this._data.has(key);
-  }
-
-  delete(key: string): void {
-    this._data.delete(key);
-  }
-
-  clear(): void {
-    this._data.clear();
-  }
-
-  toHash(): Record<string, unknown> {
-    return Object.fromEntries(this._data);
-  }
-
-  toH(): Record<string, unknown> {
-    return this.toHash();
-  }
-
-  toObject(): Record<string, unknown> {
-    return this.toHash();
-  }
-}
-
 function formatToMime(format: string): string {
   const MIMES: Record<string, string> = {
     json: "application/json",
