+++ packages/actionpack/src/action-dispatch/testing/index.ts
@@ -1,1 +1,4 @@
-export {};
+export { IntegrationTest, type IntegrationRequestOptions } from "./integration.js";
+export { TestRequest } from "./test-request.js";
+export { LiveTestResponse } from "./test-response.js";
+export { TestSession } from "./test-process.js";
