+++ packages/actionpack/src/action-dispatch/testing/test-request.ts
@@ -1,0 +1,14 @@
+import { Request } from "../http/request.js";
+
+export class TestRequest extends Request {
+  constructor(env: import("@blazetrails/rack").RackEnv = {}) {
+    super({
+      REQUEST_METHOD: "GET",
+      PATH_INFO: "/",
+      HTTP_HOST: "test.host",
+      SERVER_NAME: "test.host",
+      SERVER_PORT: "80",
+      ...env,
+    });
+  }
+}
