+++ packages/actionpack/src/action-dispatch/testing/test-response.ts
@@ -1,0 +1,3 @@
+import { Response } from "../http/response.js";
+
+export class LiveTestResponse extends Response {}
